requirements.txt
streamlit
streamlit-chat
audio-recorder-streamlit
SpeechRecognition
gTTS
langdetect
pymongo
pyaudio
uuid
google-generativeai
python-dotenv
